rename feature_astronaut to feature_astronaut_presentation
