+++ .idea/gradle.xml
@@ -11,7 +11,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/feature_astronaut" />
+            <option value="$PROJECT_DIR$/feature_astronaut_presentation" />
           </set>
         </option>
       </GradleProjectSettings>
