+++ feature_astronaut_presentation/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.alphanication.feature_astronaut"
+        applicationId "com.alphanication.feature_astronaut_presentation"
         minSdk 21
         targetSdk 32
         versionCode 1
