+++ feature_astronaut_presentation/src/main/AndroidManifest.xml
