+++ feature_astronaut_presentation/src/main/java/com/alphanication/feature_astronaut_presentation/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.alphanication.feature_astronaut
+package com.alphanication.feature_astronaut_presentation
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
