+++ feature_astronaut_presentation/src/main/res/drawable/ic_launcher_background.xml
