+++ feature_astronaut_presentation/src/main/res/layout/activity_main.xml
