+++ feature_astronaut_presentation/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml
