+++ feature_astronaut_presentation/src/main/res/values/strings.xml
