+++ settings.gradle
@@ -14,4 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "Spacestar"
 include ':app'
-include ':feature_astronaut'
+include ':feature_astronaut_presentation'
